async-llvm(9): Move OngoingCrateTranslation into back::write.
